RabbitMQ publisher for trip data from simulation (#12)
* WIP: Data extractor skeleton and setup EasyNetQ

* Docker compose for RabbitMQ

* Rebase, cleanup, renaming

* Implemented TripDataPublisher

Including dtos.
Weather and Node attributes are WIP

* Updated WeatherDataGenerator to match domain model attribute types.

* Extended TripDataItemDto with weather and edge attributes.

+++ OpenWeather/OpenWeatherService.cs
@@ -37,7 +37,7 @@ namespace OpenWeather
                     fog: CalculateFog(openWeatherResponse.Visibility, openWeatherResponse.Weather.Select(w => w.Id).ToArray()),
                     humidity: openWeatherResponse.Main.Humidity,
                     temperature: (int)openWeatherResponse.Main.Temp,
-                    precipitation: openWeatherResponse.Rain?.ThreeHours ?? 0
+                    precipitation: (int)(openWeatherResponse.Rain?.ThreeHours ?? 0)
                 );
         }
